refactor: 패키지 이동 및 변수명 변경
